separate some functionalites from the index.js
