+++ src/modules/localStorage.js
@@ -1,0 +1,6 @@
+const getDataFromLocalStorage = () => {
+  const todoList = localStorage.getItem('todoList');
+  return JSON.parse(todoList);
+};
+
+export default getDataFromLocalStorage;
